Remove and add words
Some of the removed words are not necessarily 'offensive' (and in one
case it is based on a discredited etymology) and it's possible I got rid
of some by accident. Others might not be offensive as such but have
words in them that some might consider offensive. Many (most if not all)
I do not find offensive but some might not like some of them.

I also added some obscure words (some archaic words, some not).