Add make test script and workflow
This will hopefully flag the case where someone forgets to run make sort
prior to committing.

Of course it is possible that there is a syntax error in the yml files
and if so I will fix those (though they are copies from mkiocccentry
modified for this repo).

The test.sh is very simple: create a temp words list, sort the words
list into it (like the make sort rule does) and then compare it to the
words list itself: if they differ it's an error, otherwise it is success
(exit 0).

If it is different it suggests make sort and then exists non-zero.