Remove 'human' :-)
Okay so it might not be a rude word as such but it was too good to not
do and it is not like humans are polite :-) (okay many are but the human
species is not on a whole polite and depending on the definition of rude
they, including myself, are very rude :-) . I did keep other 'offensive'
words like humanise and such but only because removing those would be
removing too many words without adding more.

Yes this is totally a joke and since it's only one word it is I deem
worth doing for the laughs (though I do realise the humour impaired
might not find it funny).